split into files

+++ run.sh
@@ -1,5 +1,9 @@
-gcc -O3 main.c -o tach.exe -lgc
-python3 main.py -ec $1
-./tach.exe
-rm out.bytecode
-rm tach.exe
+if gcc -O3 *.c -o tach.exe -lgc
+then
+    if python3 main.py -ec $1
+    then
+        time ./tach.exe
+        rm out.bytecode
+    fi
+    rm tach.exe
+fi
